Extract suggested articles component into a single component
