+++ src/app/(frontend)/literature/page.tsx
@@ -5,7 +5,7 @@ import RightMenuContainer from '@/components/RightMenuContainer'
 import BottomMenu from '@/components/BottomMenu'
 import { getCachedGlobal } from '@/utilities/getGlobals'
 import { Literature as LiteratureGlobalType } from '@/payload-types'
-import SuggestedArticles from '@/components/Literature/SuggestedArticles'
+import SuggestedArticles from '@/components/SuggestedArticles'
 import TopMenuContainer from '@/components/Literature/TopMenuContainer'
 import LiteratureContent from '@/components/Literature/LiteratureContent'
 import LiteratureHero from '@/components/Literature/LiteratureHero'
