+++ src/components/Literature/types.ts
@@ -1,9 +1,5 @@
 import { Literature } from '@/payload-types'
 
-export interface SuggestedArticlesProps {
-  articles: Literature['suggestedArticles']
-}
-
 export interface LiteratureHeroProps {
   ads: Literature['pageAds']
 }
