+++ src/components/Scope/types.ts
@@ -1,4 +1,4 @@
-import { Category, Media, Scope } from '@/payload-types'
+import { Category, Media } from '@/payload-types'
 
 interface ScopeCategory {
   id: string
@@ -23,10 +23,6 @@ export interface ScopeContentItemProps {
   title: string
   slug: string
   media: Media
-}
-
-export interface SuggestedArticlesProps {
-  articles: Scope['suggestedArticles']
 }
 
 export interface TopMenuContainerProps {
