+++ src/components/SuggestedArticles.tsx
@@ -1,15 +1,21 @@
 import Link from 'next/link'
 import React from 'react'
-import { SuggestedArticlesProps } from './types'
+import { Article } from '@/payload-types'
 
-const SuggestedArticles: React.FC<SuggestedArticlesProps> = ({ articles }) => {
+interface SuggestedArticlesProps {
+  articles: (string | Article)[] | null | undefined
+}
+
+const SuggestedArticles: React.FC<SuggestedArticlesProps> = (props) => {
+  const { articles } = props
+
   const hasArticles = articles && Array.isArray(articles) && articles.length > 0
 
   return (
-    <div className="mb-2 overflow-auto bg-black lg:mb-0">
+    <div className="overflow-auto bg-black">
       <div className="container flex w-full gap-4 overflow-x-auto py-1">
         {hasArticles &&
-          articles?.map((article) => {
+          articles.map((article) => {
             if (typeof article === 'object' && article !== null) {
               const { id, title, slug } = article
